BACKLOG-20266: Rename open in live action
