+++ src/javascript/JContent/actions/openInAction/index.js
@@ -1,0 +1,1 @@
+export * from './openInAction';
