--- src/javascript/JContent/actions/openInLiveAction/index.js
@@ -1,1 +1,0 @@
-export * from './openInLiveAction';
